Require .ORIG at start and .END at end of asm! macro
This simplifies the macro body by quite a lot, and solves several
problems:

- We no longer have to deal with the possibility of putting ORIG or END
in the middle of the program, or forgetting them; that's now not
syntactical.

- We can set the origin and the PC right at the start of the macro body
instead of conditionally emitting code to do so.

- We can always push an instruction into the vector instead of having
the sub-rule expand to an Option and then mapping over it.

- We can have the initial dot in front of the directives. This is nice
because .ORIG and .END are the ones that don't emit any bytecodes. In
the original LC-3 assembly language, .BLKW and .STRINGZ have dots too,
but they actually do translate to bytecodes.

- We can no longer have a label in front of ORIG, which makes sense
because what address should it be? Zero is probably not what you want.