Refactor assembler to assemble right in the asm macro
Instead of the asm macro returning an object which we still have to call
assemble() on, have it assemble the program already, and make assemble()
into the constructor method of Program. This changes the internal
representation of an assembled Program to store the assembled bytecode.

It also refactors the assemble() method to return a custom Error
implementation on failure, containing a list of errors in the assembly
language code, instead of returning the list of errors alongside the
bytecode. In order to easily combine the new AssemblerError with
io::Error in Result types, we add a dependency on the anyhow library.