Use bare identifiers for labels
Instead of using string literals for labels, use bare identifiers with a
colon before instead of after, to eliminate ambiguity with the opcode's
bare identifier. This saves a bit of horizontal space.

Unfortunately we can't also use bare identifiers in the arguments to the
opcodes. ($t:expr) apparently makes the token opaque so it can't then be
passed to one of the sub-rules and parsed as an ident. (Although why
that still works for literals, I'm not sure.)